requirements.txt
torch==2.0.0
torchvision==0.15.1
Pillow==9.1.1
requests==2.28.1
qai_hub
